Add function templates for immut-backed reify objs
It's not done yet, but the framework is in with this change.
Soon, these function templates should be used to generate
the TypeScript types, and that will make for easy handoffs
to C++ code.

Also, clean up the Haskell code generation output so that
there's only one interface generator file that generates
everything.