Frame Reify as a library generator.
The 3D geometry building frontend is now renamed to "hypo", and
"reify" now represents the framework for the V8 interface glue.

In other words, the input to reify is an interface, and the output
is a C++ library that can parse a TypeScript environment with the
defined interface types available, and return appropriate C++
types.