Clean up handling of hypo TypeScript module.
Make it so that the hypo TypeScript module is more consistent
between how the generated TypeScript source files are stored on
disk and how they are baked into the executable.

This makes it easier to decide later between generating the
declaration files from the binary or offline.

Also, enable the option to output declaration files.